+++ Sources/Fluent/ModelUserToken.swift
@@ -1,51 +1,5 @@
 import Vapor
 
-public protocol ModelUser: Model, Authenticatable {
-    static var usernameKey: KeyPath<Self, Field<String>> { get }
-    static var passwordHashKey: KeyPath<Self, Field<String>> { get }
-    func verify(password: String) throws -> Bool
-}
-
-extension ModelUser {
-    public static func authenticator(database: DatabaseID? = nil) -> ModelUserAuthenticator<Self> {
-        ModelUserAuthenticator<Self>(database: database)
-    }
-
-    var _$username: Field<String> {
-        self[keyPath: Self.usernameKey]
-    }
-
-    var _$passwordHash: Field<String> {
-        self[keyPath: Self.passwordHashKey]
-    }
-}
-
-public struct ModelUserAuthenticator<User>: BasicAuthenticator
-    where User: ModelUser
-{
-    public let database: DatabaseID?
-
-    public func authenticate(
-        basic: BasicAuthorization,
-        for request: Request
-    ) -> EventLoopFuture<User?> {
-        User.query(on: request.db(self.database))
-            .filter(\._$username == basic.username)
-            .first()
-            .flatMapThrowing
-        {
-            guard let user = $0 else {
-                return nil
-            }
-            guard try user.verify(password: basic.password) else {
-                return nil
-            }
-            return user
-        }
-    }
-}
-
-
 public protocol ModelUserToken: Model {
     associatedtype User: Model & Authenticatable
     static var valueKey: KeyPath<Self, Field<String>> { get }
@@ -73,7 +27,6 @@ public struct ModelUserTokenAuthenticator<Token>: BearerAuthenticator
     where Token: ModelUserToken
 {
     public typealias User = Token.User
-
     public let database: DatabaseID?
 
     public func authenticate(
@@ -97,3 +50,4 @@ public struct ModelUserTokenAuthenticator<Token>: BearerAuthenticator
         }
     }
 }
+
